backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
celery
redis
anthropic
openai
python-dotenv
alembic
